fix warn

+++ source/ShaderWeaver.cpp
@@ -753,7 +753,7 @@ sw::NodePtr ShaderWeaver::CreateWeaverNode(const bp::Node& node)
     for (int i = 0, n = node.GetAllInput().size(); i < n; ++i)
     {
         auto& imports = dst->GetImports();
-        if (node.IsExtensibleInputPorts() && i >= imports.size()) {
+        if (node.IsExtensibleInputPorts() && i >= static_cast<int>(imports.size())) {
             continue;
         }
         if (!imports[i].var.IsDefaultInput()) {
